refactor(collector): simplify interfaces and add metric registration
Address review comments from @lionelvillard and @asm582 on PR #416:

Source interface simplifications:
- Remove health-related types (SourceHealthStatus, SourceHealth, Health())
- Remove MetricCategory and SupportedCategories() - defer until needed
- Remove PushableMetricSource and MetricPushHandler - defer to later PR
- Remove Close() - sources can implement io.Closer if needed
- Update CollectionInterval comment to say 'actual' not 'recommended'

Consumer-driven metric registration:
- Add MetricRequest type for consumers to specify needed metrics
- Add Register(metric) bool - source returns true if it can provide
- Sources store registered metrics internally
- Collect() returns values for all registered metrics

Package reorganization:
- Create internal/metricscache package for cache types and interfaces
- internal/metricscache/types.go: TimeSeries, DataPoint, MetricValue, etc.
- internal/metricscache/interfaces.go: Reader, Writer, ReadWriter interfaces

The MetricSource interface now has 4 methods:
- Name() string
- CollectionInterval() time.Duration
- Register(metric MetricRequest) bool
- Collect(ctx) ([]metricscache.MetricValue, error)
