+++ internal/metricscache/types.go
@@ -14,7 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package collector
+// Package metricscache provides metrics caching for the autoscaler.
+package metricscache
 
 import (
 	"fmt"
